Admin dashboard

+++ Brickwell/Brickwell/Models/Product.cs
@@ -9,7 +9,7 @@ namespace Brickwell.Models
         public required string name { get; set; }
         public required int year { get; set; }
         public required int num_parts { get; set; }
-        public required decimal price { get; set; }
+        public required int price { get; set; }
         public required string img_link { get; set; }
         public required string primary_color { get; set; }
         public string? secondary_color { get; set; }
